Other clean up: remove imports; unnecessary boxing; Use String builder where necessary; Unnecessary semicolon; Unnecessary modifier in interfaces; Catch block may ignore exception; Stream API call chain can be simplified

+++ IGRP/src/main/java/nosi/core/webapp/import_export_v2/exports/ExportSqlHelper.java
@@ -29,7 +29,7 @@ public class ExportSqlHelper {
 		this.application = Core.findApplicationById(model.getApplication_id());
 		this.showAllTable(view,false);
 		for(String type:opcoes) {
-			int t = Core.toInt(type).intValue();
+			int t = Core.toInt(type);
 			if(t==OptionsImportExport.BPMN.getValor()) {
 				view.table_bpmn.setVisible(true);
 				this.loadBPMNData(model);
